Found 3 vulnerabilities (1 moderate, 2 high) in 24 scanned packages
Error:
- Gets on GitHub dependabot alerts: 'glob-parent', 'nth-check', 'semver-regex', 'trim-newlines'
Package│markdown-it│Moderate│Uncontrolled Resource Consumption in markdown-it│Dependency of jsdoc | Patched in >=12.3.2
Package│marked│High│Inefficient Regular Expression Complexity in marked│Dependency of jsdoc | Patched in >=4.0.10
- Running the command 'npm audit' in the terminal returns a list with 'Some vulnerabilities require your attention to resolve’.

References:
- How to override a nested npm sub-dependency with a different package altogether (not just different package version number)
https://stackoverflow.com/a/69591894
- 'npm-force-resolutions' not working when installing a new package
https://stackoverflow.com/a/67446313

Solution:
- Added each vulnerability of the dependency of dependency to the 'resolutions’ in the ‘package.json’ file.
Each vulnerability is replaced with a different package or with a more stable version of the package.
- Created npm 'preinstall' script to fix the vulnerabilities even if the 'package-lock.json' file has been deleted or does not exist yet.
This command first does the normal npm installation to create the 'package-lock.json' file and then updates the dependency resolution to the mentioned version.